requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.52.4
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
msgpack==1.0.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
regex==2024.5.15
requests==2.32.2
requests-oauthlib==1.3.1
scikit-learn==1.5.0
scipy==1.13.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
tweepy==4.14.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
urllib3==2.2.1
zope.interface==6.4.post2
